% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{apply_labels}
\alias{apply_labels}
\title{Apply Labels from a Definition Data Frame}
\usage{
apply_labels(data, label_def)
}
\arguments{
\item{data}{Data frame to apply labels}

\item{label_def}{Data frame with columns: variable, label, value_labels}
}
\value{
Data frame with labels applied
}
\description{
Apply Labels from a Definition Data Frame
}
\examples{
# Define labels
label_def <- data.frame(
  variable = c("gender", "age"),
  label = c("Gender", "Age in years"),
  value_labels = c("1=Male; 2=Female", NA)
)

# Apply labels
df <- apply_labels(df, label_def)

}
