% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{apply_spss_labels}
\alias{apply_spss_labels}
\title{Apply SPSS Label Syntax to Data Frame}
\usage{
apply_spss_labels(data, syntax = NULL, file = NULL)
}
\arguments{
\item{data}{Data frame to apply labels}

\item{syntax}{Character string containing SPSS syntax. If NULL, \code{file} must be specified.}

\item{file}{Path to an SPSS syntax file (.sps). If NULL, \code{syntax} must be specified.}
}
\value{
Data frame with labels applied
}
\description{
Parse SPSS VARIABLE LABELS and VALUE LABELS syntax and apply them to a data frame.
Combines \code{\link{parse_spss_labels}} and \code{\link{apply_labels}}.
}
\examples{
spss_syntax <- '
VARIABLE LABELS
  gender "Gender"
  age "Age in years".
VALUE LABELS
  gender
    1 "Male"
    2 "Female".
'
df <- apply_spss_labels(df, spss_syntax)

# From .sps file
# df <- apply_spss_labels(df, file = "labels.sps")

}
