% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occ_scale.R
\name{check_occ_scale}
\alias{check_occ_scale}
\title{Check Unmatched Occupation Codes}
\usage{
check_occ_scale(data, code, type = c("ssm", "jess"))
}
\arguments{
\item{data}{Data frame}

\item{code}{Variable name containing occupation codes (unquoted)}

\item{type}{Scale type: "ssm" for SSM occupation codes (default),
"jess" for JESS 232 occupation codes}
}
\value{
Tibble with columns \code{code} (unmatched code values) and \code{n}
(number of observations). Returns empty tibble if all codes match.
}
\description{
Identify occupation codes in the data that do not have corresponding
SEI/SSI scale values.
}
\examples{
df <- data.frame(occ = c(501, 502, 999, 999))
check_occ_scale(df, occ, type = "ssm")

}
