% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occ_scale.R
\name{occ_scale}
\alias{occ_scale}
\title{Assign Occupational Scales (SEI/SSI)}
\source{
Data from OccupationalScales v1.0:
\url{https://github.com/ShoFujihara/OccupationalScales}
}
\usage{
occ_scale(data, code, type = c("ssm", "jess"), quiet = FALSE)
}
\arguments{
\item{data}{Data frame}

\item{code}{Variable name containing occupation codes (unquoted)}

\item{type}{Scale type: "ssm" for SSM occupation codes (default),
"jess" for JESS 232 occupation codes}

\item{quiet}{If TRUE, suppress messages about unmatched codes (default = FALSE)}
}
\value{
Data frame with \code{.sei} and \code{.ssi} columns added.
Unmatched codes will have \code{NA} for both columns.
}
\description{
Look up Socio-Economic Index (SEI) and Social Status Index (SSI) values
for occupation codes and add them as new columns to the data frame.
}
\examples{
df <- data.frame(occ = c(501, 502, 999))
occ_scale(df, occ, type = "ssm")

# With dplyr pipe
# data |> occ_scale(occupation, type = "ssm")

}
