% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{parse_spss_labels}
\alias{parse_spss_labels}
\title{Parse SPSS Label Syntax}
\usage{
parse_spss_labels(syntax = NULL, file = NULL)
}
\arguments{
\item{syntax}{Character string containing SPSS syntax. If NULL, \code{file} must be specified.}

\item{file}{Path to an SPSS syntax file (.sps). If NULL, \code{syntax} must be specified.}
}
\value{
Data frame with columns: variable, label, value_labels
}
\description{
Parse SPSS VARIABLE LABELS and VALUE LABELS syntax into a data frame
compatible with \code{\link{apply_labels}}.
}
\examples{
spss_syntax <- '
VARIABLE LABELS
  gender "Gender"
  age "Age in years".
VALUE LABELS
  gender
    1 "Male"
    2 "Female".
'
label_def <- parse_spss_labels(spss_syntax)

}
