% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/socialsurvey-package.R
\docType{package}
\name{socialsurvey-package}
\alias{socialsurvey}
\alias{socialsurvey-package}
\title{socialsurvey: R Functions for Social Survey Data Analysis}
\description{
A collection of R functions for social survey data analysis. Includes tools for frequency tables, metadata management, variable labels, value labels, and codebook generation. Supports Stata, SPSS, and SAS files.
}
\author{
\strong{Maintainer}: Sho Fujihara \email{sho.fujihara@iss.u-tokyo.ac.jp}

}
\keyword{internal}
