% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{set_val_labels}
\alias{set_val_labels}
\title{Set Value Labels}
\usage{
set_val_labels(data, var, ...)
}
\arguments{
\item{data}{Data frame}

\item{var}{Variable name (unquoted)}

\item{...}{Value-label pairs (e.g., \code{1} = "Male", \code{2} = "Female")}
}
\value{
Data frame with value labels applied
}
\description{
Set Value Labels
}
\examples{
df <- set_val_labels(df, gender, `1` = "Male", `2` = "Female")

}
