% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{set_var_label}
\alias{set_var_label}
\title{Set Variable Label}
\usage{
set_var_label(data, var, label)
}
\arguments{
\item{data}{Data frame}

\item{var}{Variable name (unquoted)}

\item{label}{Label string}
}
\value{
Data frame with label applied
}
\description{
Set Variable Label
}
\examples{
df <- data.frame(gender = c(1, 2, 1), age = c(25, 30, 35))
df <- set_var_label(df, gender, "Gender")

}
